+++ tools/vf-extensions/gulp-tasks/gulp-build-search-index.js
@@ -8,7 +8,10 @@
  *
  * Results in a search index at:
  *  buildDestionation+'/search_index.js'
- * Requires `yarn add strip-js striptags`
+ *
+ * Requires `yarn add strip-js striptags node-html-parser`
+ *
+ * Elements wrapped with `class="vf-search-client-side--no-index"` will not be logged in the index
  */
 module.exports = function(gulp, path, buildDestionation) {
 
@@ -17,6 +20,7 @@ module.exports = function(gulp, path, buildDestionation) {
   const through = require('through2');
   const stripJs = require('strip-js');
   const striptags = require('striptags');
+  const HTMLParser = require('node-html-parser'); // https://www.npmjs.com/package/node-html-parser
 
   gulp.task('vf-build-search-index', function() {
     const fileName = buildDestionation+'/search_index.js';
@@ -33,9 +37,19 @@ module.exports = function(gulp, path, buildDestionation) {
 
       let title = text.match(/<title>(.*?)<\/title>/gi) + ' ';
           title = title.replace(/<title>(.*?)<\/title>/gi, '$1');
-          title = title.replace('| EMBL Design System | EMBL','');
+          title = title.split('|')[0]; // only keep anything before the first pipe
 
       let body = text.match(/<body.[\s\S]*?>(.[\s\S]*?)body>/gi) + ' ';
+
+      // remove any elements with class="vf-search-client-side--no-index"
+      let bodyHtml = HTMLParser.parse(body);
+          bodyHtml.querySelectorAll('.vf-search-client-side--no-index').forEach(function(a){
+            a.set_content('');
+          })
+          body = new String(bodyHtml);
+
+
+
           body = stripJs(striptags(body));
           body = body.replace(/&quot;/g, ' '); // remove white space
           body = body.replace(/class\=/g, ' '); // remove white space
@@ -46,6 +60,7 @@ module.exports = function(gulp, path, buildDestionation) {
           body = body.replace(/   /g, ' '); // remove white space
           body = body.replace(/  /g, ' '); // remove white space
           body = body.replace(/"/g, '\''); // remove double quotes
+
 
       output += endOfLine + '{"id":"'+counter+'", "title": "'+title+'", "text": "'+body+'", "tags": "", ';
       counter = counter + 1;
@@ -60,7 +75,7 @@ module.exports = function(gulp, path, buildDestionation) {
       cb(null, file)
     })
     .on('finish', function (status) {
-      gutil.log(gutil.colors.green('Finished prepping JSON'));
+      gutil.log(gutil.colors.green('Finished prepping search JSON'));
         // write the rendered JSON
         fs.writeFileSync(fileName, output + endOfLine + ']};');
       })