+++ tools/vf-extensions/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.0.0-alpha.9
 
 * Update documentation note on pointing to `gulp-build-search-index.js`
+* Search indexer now removes html areas wrapped in `class="vf-search-client-side--no-index"`